cpplint

+++ Application/tasks.c
@@ -16,24 +16,27 @@
 /* Although normally "The Standard Library input/output functions shall not be
    used" (Rule 21.6), in this case we make use of it as temporary solution. */
 // cppcheck-suppress misra-c2012-21.6
-#include "tasks.h"
+#include "Application/tasks.h"
 
 #include <stdint.h>
 #include <stdio.h>
 
-#include "helpers.h"
+#include "Application/helpers.h"
 
 void task1(void* arg) {
-    uint32_t computation_time = (uint32_t) * ((uint32_t*)arg);
+    uint32_t computation_time =
+        *((uint32_t*)arg);  // NOLINTNEXTLINE(readability/casting)
     busy_wait_ms(computation_time);
 }
 
 void task2(void* arg) {
-    uint32_t computation_time = (uint32_t) * ((uint32_t*)arg);
+    uint32_t computation_time =
+        *((uint32_t*)arg);  // NOLINTNEXTLINE(readability/casting)
     busy_wait_ms(computation_time);
 }
 
 void task3(void* arg) {
-    uint32_t computation_time = (uint32_t) * ((uint32_t*)arg);
+    uint32_t computation_time =
+        *((uint32_t*)arg);  // NOLINTNEXTLINE(readability/casting)
     busy_wait_ms(computation_time);
 }
